Add list view for list of children Add AddChildActivity
In the ChildListActivity, there is now a ListView which can
display a list of all the children currently added.
This ListView uses the xml file child_name_view.xml to choose the
formatting for the ListView. The child_name_view.xml can be edited
to change the ListView.
Need to add an output for when the list is empty.

Added an activity which takes the user's input for a child's name
then stores that name into the child manager.
When a new child is added the add child activity will end and
go back to the child list activity. There will also be a toast
message saying the name which was added to the manager. The child
list activity gets updated everytime it is again resumed.