Update FlipActivity Update activity_flip.xml Add heads image of coin Add tails image of coin
The FlipActivity will flip the coin with an onClickListener.
When the coin is clicked, there will be an animation,
currently the coin starts to heads and always ends on tails.

The activity_flip.xml has been updated to include an image view
for the coin, initially it is set to heads.