Add/Edit Child works perfectly
The user can now add or edit a child. The app will handle case for empty string and will also show a correct dialog box if the user decides to change anything. The user can also delete a child from his list. And, the child list will be saved between runs, so that user doesn't have to re-enter the names of the child everytime.