Add buttons for user to choose coin side Add model class to store history of flips Add model class to store information of current flip Update ChildManager
In the xml file for the flip activity, added 2 buttons to allow user to choose the side
of the coin they think will win.

2 classes have been added to the model package.
SingleFlipInformation stores the information for the current flip. Stores information such
as: child name, child's choice, time of the flip, and whether the child won.
HistoryOfFlips stores an ArrayList of flips which have been done, as well as the index
of the current child in the ChildManager.

Added a size method to the ChildManager to know when end of list is reached, stops error
for when end of list is reached and need to get name of next child.