Add listview to show history of all flips Add images for correct or incorrect coin flip guess
The list view in the FlipHistoryActivity shows the date and time of flip, name of the child
who flipped it, the guess they had and whether or not the guess was correct.
If the guess is correct it says true (need to change this to something like "winner") and
shows an image ofa check mark. If they guess was incorrect it says false and shows and x.