Update index after child deleted in AddChildActivity Update FlipActivity Update HistoryOfFlips
In AddChildActivity added a few lines to check if current child index needs to be updated after
a child has been removed.

In FlipActivity made the child index get incremented after a flip has happened. Added a textview
to show the index of the current child that flips.

Update function name in HistoryOfFlips