Done with Divij Rename SingleFlipInformation to CoinFlip Rename single_flip_information to flip_history_view Update method for storing name of child for coin flip Update what triggers coin flip Update FlipHistoryActivity
The name of the child is stored by their index in the ChildManager. This is so, even if the name gets edited its data
still stays the same and does not get deleted or lost.

Once the user chooses whether the coin will land on heads or tails, the coin gets flipped. Rather than having to flip
the coin by clicking the image, the buttons trigger the flip.

Flip history activity shows a ListView of all flips done. The appearance of it has been update to look better and easier
to understand at a quick glance.