Update to store history of flips in SharedPreferences Add ImageButton for user to clip to flip coin Update to correct some formatting
Flip history manager stores history of all flips when the coin animation
has finished.

An image button looking like a person flipping a coin has been added for the
user to click for flipping the coin. After the user estimates heads or tails,
they can click the flip coin button to start the animation.

Refactored and reformatted some of the code to make it look more consistent.
Extracted duplicate code into a method.