Update buttons to be enabled and disabled in FlipActivity
When there are no children configured the buttons to choose heads and tails
will not be enabled. Only the button to flip the coin and go the the history
will be enabled. If there are children heads, tails and history buttons will
be enabled while the flip coin button is disabled. Once the user has chosen
heads or tails, the flip coin button will be enabled.

During the coin flip animation ALL buttons will be disabled.