Fix UI
Fixed ChildListActivity and AddChildActivity's UI to auto resize with the screen size.
The only thing that isn't working properly is the edit text on the add/edit child activity. It full thing resizes properly, but the text size doesn't resize properly.
Also, for the fab on the child list activity looks smaller on the tablet, so I tried to resize it, but didn't work so I left it as it is.